requirements.txt
fastapi
uvicorn[standard]
redis>=4.0.0
python-multipart
openai
pydantic-settings
google-genai
httpx
python-dotenv
